requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-socketio==5.11.4
SQLAlchemy==2.0.36
alembic==1.14.0
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.12
psycopg2-binary==2.9.10
